+++ app/src/test/java/com/scepticalphysiologist/dmaple/map/camera/FrameRateTimerTest.kt
@@ -1,6 +1,6 @@
 // Copyright (c) 2025-2025. Dr Sean Paul Parsons. All rights reserved.
 
-package com.scepticalphysiologist.dmaple.map
+package com.scepticalphysiologist.dmaple.map.camera
 
 import com.scepticalphysiologist.dmaple.assertNumbersEqual
 import org.junit.Assert.assertEquals