+++ changeLog.md
@@ -1,5 +1,6 @@
 # DMaoLE for Android: change log
 
+- 03/06/2025: Refactor: Camera service into its own class (PR#24).
 - 02/06/2025: Feature: Added manual focus (PR#22).
 
 ### Version 1.0